FIX: 댓글 인풋 댓글 컴포넌트로 이동
